Update showcase folder
