+++ components/Header.js
@@ -10,7 +10,7 @@ const Header = (props) => (
 				<img id="logo-small" src={asset('/images/logo_mini.svg')} />
 			</Link>
 			<ul>
-				<li><Link href="/showcase" passHref><a className="menu">Projets</a></Link></li>
+				<li><Link href="/showcase/" passHref><a className="menu">Projets</a></Link></li>
 				<li><Link href="/resources" passHref><a className="menu">Ressources</a></Link></li>
 				<li><a className="menu" href="https://openfisca.org/doc/">Documentation</a></li>
 				<li><Link href="/community" passHref><a className="menu">Communauté</a></Link></li>
